+++ metrics/ttfp/benchmark-ttfp.jl
@@ -7,8 +7,17 @@ macro ctime(x)
     end
 end
 t_using = @ctime @eval using $Package
-Makie.inline!(false) # needed for cairomakie to return a screen
 
+function get_colorbuffer(fig)
+    # We need to handle old versions of Makie
+    if isdefined(Makie, :CURRENT_BACKEND) # new version after display refactor
+        return Makie.colorbuffer(fig) # easy :)
+    else
+        Makie.inline!(false)
+        screen = display(fig; visible=false)
+        return Makie.colorbuffer(screen)
+    end
+end
 
 if Package == :WGLMakie
     using ElectronDisplay
@@ -18,7 +27,7 @@ if Package == :WGLMakie
 end
 
 create_time = @ctime fig = scatter(1:4; color=1:4, colormap=:turbo, markersize=20, visible=true)
-display_time = @ctime Makie.colorbuffer(display(fig))
+display_time = @ctime get_colorbuffer(fig)
 
 using BenchmarkTools
 using BenchmarkTools.JSON
@@ -32,7 +41,7 @@ old = isfile(result) ? JSON.parse(read(result, String)) : [[], [], [], [], []]
 push!.(old[1:3], [t_using, create_time, display_time])
 
 b1 = @benchmark fig = scatter(1:4; color=1:4, colormap=:turbo, markersize=20, visible=true)
-b2 = @benchmark Makie.colorbuffer(display(fig))
+b2 = @benchmark get_colorbuffer(fig)
 
 using Statistics